Moved network response handling to RandomRepository, and added fallback to get random number from local Random number generator if network call fails
